requirements.txt
# FastAPI and web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6

# Authentication and security
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Database
pymongo>=4.6.0
pinecone>=5.0.0  # Changed from pinecone-client

# LangChain and AI
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.6.0
langgraph>=0.1.0

# Data validation and settings
pydantic[email]>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# Environment and configuration
python-dotenv>=1.0.0

# HTTP client
httpx>=0.25.0
requests>=2.31.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Email support
aiosmtplib>=3.0.0

# Date and time utilities
python-dateutil>=2.8.0

# Neural networks and ML
torch>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0

# Additional utilities
typing-extensions>=4.8.0
mistralai